chore: rename 'components' folder as 'scenes' and modify some import routes
