+++ src/scenes/ChallengeMode/gameConfig.js
