+++ src/scenes/ChallengeMode/hooks/useThemeManager.js
