+++ src/scenes/ChallengeMode/hooks/useZombieManager.js
