+++ src/scenes/ChallengeMode/zombie1.png
