+++ src/scenes/ChallengeMode/zombie2.png
