+++ src/scenes/ChallengeMode/zombie3.png
