+++ src/scenes/ChallengeMode/zombie4.png
