+++ src/scenes/Option/Option.css
@@ -1,4 +1,4 @@
-/* src/components/Option/Option.css */
+/* src/scenes/Option/Option.css */
 
 /* ------------------------------------------------------------ */
 /* 1. 容器樣式 */
